requirements.txt
alembic==1.7.6
autopep8==1.6.0
blinker==1.4
click==7.1.2
dominate==2.6.0
Flask==1.1.4
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Toastr==0.5.6
Flask-Uploads==0.2.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.10.1
importlib-resources==5.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2==2.9.3
pycodestyle==2.8.0
python-decouple==3.6
SQLAlchemy==1.4.31
toml==0.10.2
visitor==0.1.3
Werkzeug==0.16.0
WTForms==3.0.1
zipp==3.7.0
